Updated to target iOS 7 and create 64-bit/32-bit fat build; updated where NSInteger and int were being used interchangably - now use NSInteger exclusively as much as possible; updated icons according to new requirements; updated the plist to hide the status bar

+++ Edge Solitaire/CardSpot.m
@@ -119,7 +119,7 @@
 	[self.cardImage removeFromSuperview];
 	if(card != nil)
 	{
-		NSString *cardName = [NSString stringWithFormat:@"%@%i", card.suitFirstCharacter, card.value];
+		NSString *cardName = [NSString stringWithFormat:@"%@%li", card.suitFirstCharacter, (long)card.value];
 		self.cardImage = [[UIImageView alloc] initWithFrame:CGRectMake(0, 0, self.bounds.size.width, self.bounds.size.height)];
 		self.cardImage.image = [self imageWithName:cardName];
 		[self addSubview:self.cardImage];
